refactor: move components to another dir
