+++ src/app/(main)/login/page.tsx
@@ -1,4 +1,4 @@
-import SignIn from '../_components/signIn';
+import SignIn from '@/components/signIn';
 
 export default function SignInPage() {
   return <SignIn />;
